requirements.txt
google-genai
openai
python-dotenv
rich
pyyaml
anthropic
